Create Dec-5
https://leetcode.com/problems/move-pieces-to-obtain-a-string/?envType=daily-question&envId=2024-12-05

A few insights that will help avoid having to handle a lot of edge cases:

Because letters cannot pass each other, the i'th letter you encounter in start must also be the i'th letter you encounter in target.
If the i'th letter in start is 'L', its index in start must be greater than or equal to its index in target. The same applies for 'R', but the index in start must be less than or equal to the index in target.